update
- 路由器配置方式调整
- application server 启动运行逻辑调整
- 一些更新
