+++ app/config/base.php
@@ -14,18 +14,14 @@ return [
     ],
     'router'      => [
         'class'  => \swoft\web\Router::class,
-        'config' => [
-            'ignoreLastSep'  => false,
-            'tmpCacheNumber' => 100,
-            'matchAll'       => '',
+        'ignoreLastSep'  => false,
+        'tmpCacheNumber' => 100,
+        'matchAll'       => '',
 
-            // auto route match @like yii framework
-            'autoRoute'      => [
-                'enable'              => true,
-                'controllerNamespace' => 'app\\controllers',
-                'controllerSuffix'    => 'Controller',
-            ],
-        ]
+        // auto route match @like yii framework
+        'autoRoute'      => true,
+        'controllerNamespace' => 'app\\controllers',
+        'controllerSuffix'    => 'Controller',
     ],
 
     'commonParamsFilter' => [
