+++ app/config/define.php
@@ -1,12 +1,14 @@
 <?php
 // 系统名称
 define('SYSTEM_NAME', 'swoft');
+// 基础目录
+define('BASE_PATH', dirname(__DIR__, 2));
 // 应用目录
-define('APP_PATH', dirname(__DIR__, 2));
+define('APP_PATH', BASE_PATH . '/app');
 // 日志目录
-define('RUNTIME_PATH', APP_PATH . '/runtime/' . SYSTEM_NAME);
+define('RUNTIME_PATH', BASE_PATH . '/runtime/' . SYSTEM_NAME);
 // 视图目录
-define('VIEWS_PATH', APP_PATH . '/app/views');
+define('VIEWS_PATH', APP_PATH . '/views');
 // swoft启动配置文件
-define('SETTING_PATH', APP_PATH . '/bin/swoft.ini');
+define('SETTING_PATH', BASE_PATH . '/bin/swoft.ini');
