+++ app/routes.php
@@ -3,4 +3,5 @@ use app\controllers\IndexController;
 
 /* @var \swoft\web\Router $router */
 $router = \swoft\App::getBean('router');
+
 $router->get('/', IndexController::class);
