+++ bin/bootstrap.php
@@ -3,5 +3,6 @@ require_once __DIR__. '/../vendor/autoload.php';
 require_once __DIR__. '/../app/config/define.php';
 require_once __DIR__. '/../app/config/model.php';
 
-$config = require_once __DIR__. "/../app/config/base.php";
+$config = require dirname(__DIR__). '/app/config/base.php';
+
 $beanFactory = new \swoft\di\BeanFactory($config);
