+++ bin/swoft.php
@@ -1,10 +1,9 @@
 <?php
+use swoft\base\ApplicationContext;
+
 require_once __DIR__. '/bootstrap.php';
-
-$router = \swoft\base\ApplicationContext::getBean('router');
-
 require dirname(__DIR__) . '/app/routes.php';
 
 /* @var  \swoft\web\Application $application */
-$application = \swoft\base\ApplicationContext::getBean('application');
+$application = ApplicationContext::getBean('application');
 $application->run();
